removed toefl files in inc/dg
