+++ inc/geometries/fields.h
@@ -1,11 +1,5 @@
 #pragma once
 
-#include <iostream>
-#include <fstream>
-#include <cmath>
-#include <vector>
-
-#include "dg/blas.h"
 
 #include "dg/backend/functions.h"
 #include "dg/functors.h"
@@ -15,7 +9,9 @@
  *
  * Geometry objects 
  */
-namespace solovev
+namespace dg
+{
+namespace fields
 {
 ///@addtogroup geom
 ///@{
@@ -479,6 +475,61 @@ struct FieldT
 
 };
 
+//////////////////////////////////////////////////////////////////////////////
+/**
+ * @brief R component of magnetic field unit vector \f$ b_R\f$
+ */
+template<class Collective>
+struct BHatR
+{
+    BHatR( const Collective& c, double R0): c_(c), R_0(R0), invB_(c, R0){ }
+    double operator()( double R, double Z, double phi) const
+    {
+        return  invB_(R,Z)*R_0/R*c_.psipZ(R,Z);
+    }
+    private:
+    Collective c_;
+    double R_0;
+    InvB<Collective>   invB_;
+
+};
+/**
+ * @brief Z component of magnetic field unit vector \f$ b_Z\f$
+ */
+template<class Collective>
+struct BHatZ
+{
+    BHatZ( const Collective& c, double R0): c_(c), R_0(R0), invB_(c, R0){ }
+
+    double operator()( double R, double Z, double phi) const
+    {
+        return  -invB_(R,Z)*R_0/R*c_.psipR(R,Z);
+    }
+    private:
+    Collective c_;
+    double R_0;
+    InvB<Collective>   invB_;
+
+};
+/**
+ * @brief Phi component of magnetic field unit vector \f$ b_\Phi\f$
+ */
+template<class Collective>
+struct BHatP
+{
+    BHatP( const Collective& c, double R0): c_(c), R_0(R0), invB_(c, R0){ }
+    double operator()( double R, double Z, double phi) const
+    {
+        return invB_(R,Z)*R_0*c_.ipol(R,Z)/R/R;
+    }
+    
+    private:
+    Collective c_;
+    double R_0;
+    InvB<Collective>   invB_;
+  
+}; 
+////////////////////////////////////////////for grid generation/////////////////
 namespace flux{
 
 /**
@@ -968,571 +1019,7 @@ struct MinimalCurve
     PsiY psipZ_;
 };
 
-//////////////////////////////////////////////////////////////////////////////
-/**
- * @brief R component of magnetic field\f$ b_R\f$
- */
-template<class Collective>
-struct BHatR
-{
-    BHatR( const Collective& c, double R0): c_(c), R_0(R0), invB_(c, R0){ }
-    double operator()( double R, double Z, double phi) const
-    {
-        return  invB_(R,Z)*R_0/R*c_.psipZ(R,Z);
-    }
-    private:
-    Collective c_;
-    double R_0;
-    InvB<Collective>   invB_;
-
-};
-/**
- * @brief Z component of magnetic field \f$ b_Z\f$
- */
-template<class Collective>
-struct BHatZ
-{
-    BHatZ( const Collective& c, double R0): c_(c), R_0(R0), invB_(c, R0){ }
-
-    double operator()( double R, double Z, double phi) const
-    {
-        return  -invB_(R,Z)*R_0/R*c_.psipR(R,Z);
-    }
-    private:
-    Collective c_;
-    double R_0;
-    InvB<Collective>   invB_;
-
-};
-/**
- * @brief Phi component of magnetic field \f$ b_\Phi\f$
- */
-template<class Collective>
-struct BHatP
-{
-    BHatP( const Collective& c, double R0): c_(c), R_0(R0), invB_(c, R0){ }
-    double operator()( double R, double Z, double phi) const
-    {
-        return invB_(R,Z)*R_0*c_.ipol(R,Z)/R/R;
-    }
-    
-    private:
-    Collective c_;
-    double R_0;
-    InvB<Collective>   invB_;
-  
-}; 
 ///@} 
-///@addtogroup profiles
-///@{
+} //namespace fields
+} //namespace dg
 
-/**
- * @brief Delta function for poloidal flux \f$ B_Z\f$
-     \f[ |\nabla \psi_p|\delta(\psi_p(R,Z)-\psi_0) = \frac{\sqrt{ (\nabla \psi_p)^2}}{\sqrt{2\pi\varepsilon}} \exp\left(-\frac{(\psi_p(R,Z) - \psi_{0})^2}{2\varepsilon} \right)  \f]
- */
-template<class Collective>
-struct DeltaFunction
-{
-    DeltaFunction(const Collective& c, double epsilon,double psivalue) :
-        c_(c),
-        epsilon_(epsilon),
-        psivalue_(psivalue){
-    }
-    /**
-    * @brief Set a new \f$ \varepsilon\f$
-    *
-    * @param eps new value
-    */
-    void setepsilon(double eps ){epsilon_ = eps;}
-    /**
-    * @brief Set a new \f$ \psi_0\f$
-    *
-    * @param psi_0 new value
-    */
-    void setpsi(double psi_0 ){psivalue_ = psi_0;}
-
-    /**
-     *@brief \f[ \frac{\sqrt{ (\nabla \psi_p)^2}}{\sqrt{2\pi\varepsilon}} \exp\left(-\frac{(\psi_p(R,Z) - \psi_{0})^2}{2\varepsilon} \right)  \f]
-     */
-    double operator()( double R, double Z) const
-    {
-        double psip = c_.psip(R,Z), psipR = c_.psipR(R,Z), psipZ = c_.psipZ(R,Z);
-        return 1./sqrt(2.*M_PI*epsilon_)*
-               exp(-( (psip-psivalue_)* (psip-psivalue_))/2./epsilon_)*sqrt(psipR*psipR +psipZ*psipZ);
-    }
-    /**
-     * @brief == operator()(R,Z)
-     */ 
-    double operator()( double R, double Z, double phi) const
-    {
-        return (*this)(R,Z);
-    }
-    private:
-    Collective c_;
-    double epsilon_;
-    double psivalue_;
-};
-
-/**
- * @brief Flux surface average over quantity
- *
- * The average is computed using the formula
- \f[ <f>(\psi_0) = \frac{1}{A} \int dV \delta(\psi_p(R,Z)-\psi_0) |\nabla\psi_p|f(R,Z) \f]
- with \f$ A = \int dV \delta(\psi_p(R,Z)-\psi_0)|\nabla\psi_p|\f$
- * @tparam container  The container class of the vector to average
-
- */
-template <class Collective, class container = thrust::host_vector<double> >
-struct FluxSurfaceAverage
-{
-     /**
-     * @brief Construct from a field and a grid
-     * @param g2d 2d grid
-     * @param gp  geometry parameters
-     * @param f container for global safety factor
-     */
-    FluxSurfaceAverage(const dg::Grid2d& g2d, const Collective& c, const container& f) :
-    g2d_(g2d),
-    f_(f),
-    deltaf_(DeltaFunction<Collective>(c,0.0,0.0)),
-    w2d_ ( dg::create::weights( g2d_)),
-    oneongrid_(dg::evaluate(dg::one,g2d_))              
-    {
-        dg::HVec psipRog2d  = dg::evaluate( c.psipR, g2d_);
-        dg::HVec psipZog2d  = dg::evaluate( c.psipZ, g2d_);
-        double psipRmax = (double)thrust::reduce( psipRog2d.begin(), psipRog2d.end(),  0.,     thrust::maximum<double>()  );    
-        //double psipRmin = (double)thrust::reduce( psipRog2d.begin(), psipRog2d.end(),  psipRmax,thrust::minimum<double>()  );
-        double psipZmax = (double)thrust::reduce( psipZog2d.begin(), psipZog2d.end(), 0.,      thrust::maximum<double>()  );    
-        //double psipZmin = (double)thrust::reduce( psipZog2d.begin(), psipZog2d.end(), psipZmax,thrust::minimum<double>()  );   
-        double deltapsi = fabs(psipZmax/g2d_.Ny()/g2d_.n() +psipRmax/g2d_.Nx()/g2d_.n());
-        //deltaf_.setepsilon(deltapsi/4.);
-        deltaf_.setepsilon(deltapsi); //macht weniger Zacken
-    }
-    /**
-     * @brief Calculate the Flux Surface Average
-     *
-     * @param psip0 the actual psi value for q(psi)
-     */
-    double operator()(double psip0)
-    {
-        deltaf_.setpsi( psip0);
-        container deltafog2d = dg::evaluate( deltaf_, g2d_);    
-        double psipcut = dg::blas2::dot( f_,w2d_,deltafog2d); //int deltaf psip
-        double vol     = dg::blas2::dot( oneongrid_ , w2d_,deltafog2d); //int deltaf
-        double fsa = psipcut/vol;
-        return fsa;
-    }
-    private:
-    dg::Grid2d g2d_;
-    container f_;
-    DeltaFunction<Collective> deltaf_;    
-    const container w2d_;
-    const container oneongrid_;
-};
-/**
- * @brief Class for the evaluation of the safety factor q
- * \f[ q(\psi_0) = \frac{1}{2\pi} \int dV |\nabla\psi_p| \delta(\psi_p-\psi_0) \alpha( R,Z) \f]
- * @tparam container 
- *
- */
-template <class Collective, class container = thrust::host_vector<double> >
-struct SafetyFactor
-{
-     /**
-     * @brief Construct from a field and a grid
-     * @param g2d 2d grid
-     * @param gp  geometry parameters
-     * @param f container for global safety factor
-     */
-    SafetyFactor(const dg::Grid2d& g2d, const Collective& c, const container& f) :
-    g2d_(g2d),
-    f_(f), //why not directly use Alpha??
-    deltaf_(DeltaFunction<Collective>(c,0.0,0.0)),
-    w2d_ ( dg::create::weights( g2d_)),
-    oneongrid_(dg::evaluate(dg::one,g2d_))              
-    {
-      dg::HVec psipRog2d  = dg::evaluate( c.psipR, g2d_);
-      dg::HVec psipZog2d  = dg::evaluate( c.psipZ, g2d_);
-      double psipRmax = (double)thrust::reduce( psipRog2d.begin(), psipRog2d.end(), 0.,     thrust::maximum<double>()  );    
-      //double psipRmin = (double)thrust::reduce( psipRog2d.begin(), psipRog2d.end(),  psipRmax,thrust::minimum<double>()  );
-      double psipZmax = (double)thrust::reduce( psipZog2d.begin(), psipZog2d.end(), 0.,      thrust::maximum<double>()  );    
-      //double psipZmin = (double)thrust::reduce( psipZog2d.begin(), psipZog2d.end(), psipZmax,thrust::minimum<double>()  );   
-      double deltapsi = fabs(psipZmax/g2d_.Ny() +psipRmax/g2d_.Nx());
-      //deltaf_.setepsilon(deltapsi/4.);
-      deltaf_.setepsilon(4.*deltapsi); //macht weniger Zacken
-    }
-    /**
-     * @brief Calculate the q profile over the function f which has to be the global safety factor
-     * \f[ q(\psi_0) = \frac{1}{2\pi} \int dV |\nabla\psi_p| \delta(\psi_p-\psi_0) \alpha( R,Z) \f]
-     *
-     * @param psip0 the actual psi value for q(psi)
-     */
-    double operator()(double psip0)
-    {
-        deltaf_.setpsi( psip0);
-        container deltafog2d = dg::evaluate( deltaf_, g2d_);    
-        double q = dg::blas2::dot( f_,w2d_,deltafog2d)/(2.*M_PI);
-        return q;
-    }
-    private:
-    dg::Grid2d g2d_;
-    container f_;
-    DeltaFunction<Collective> deltaf_;    
-    const container w2d_;
-    const container oneongrid_;
-};
-/**
- * @brief Global safety factor
-\f[ \alpha(R,Z) = \frac{|B^\varphi|}{R|B^\eta|} = \frac{I_{pol}(R,Z)}{R|\nabla\psi_p|} \f]
- */
-template<class Collective>
-struct Alpha
-{
-    Alpha( const Collective& c):c_(c){}
-
-    /**
-    * @brief \f[ \frac{ I_{pol}(R,Z)}{R \sqrt{\nabla\psi_p}} \f]
-    */
-    double operator()( double R, double Z) const
-    {
-        double psipR = c_.psipR(R,Z), psipZ = c_.psipZ(R,Z);
-        return (1./R)*(c_.ipol(R,Z)/sqrt(psipR*psipR + psipZ*psipZ )) ;
-    }
-    /**
-     * @brief == operator()(R,Z)
-     */ 
-    double operator()( double R, double Z, double phi) const
-    {
-        return operator()(R,Z);
-    }
-    private:
-    Collective c_;
-};
-
-template<class Collective>
-struct FuncNeu
-{
-    FuncNeu( const Collective& c):c_(c){}
-    double operator()(double R, double Z, double phi) const {return -c_.psip(R,Z)*cos(phi);
-    }
-    private:
-    Collective c_;
-};
-template<class Collective>
-struct DeriNeu
-{
-    DeriNeu( const Collective& c, double R0):c_(c), bhat_(c, R0){}
-    double operator()(double R, double Z, double phi) const {return c_.psip(R,Z)*bhat_(R,Z,phi)*sin(phi);
-    }
-    private:
-    Collective c_;
-    BHatP<Collective> bhat_;
-};
-
-//psi * cos(theta)
-template<class Collective>
-struct FuncDirPer
-{
-    FuncDirPer( const Collective& c, double R0, double psi_0, double psi_1, double k):
-        R_0_(R0), psi0_(psi_0), psi1_(psi_1), k_(k), c_(c) {}
-    double operator()(double R, double Z) const {
-        double psip = c_.psip(R,Z);
-        double result = (psip-psi0_)*(psip-psi1_)*cos(k_*theta(R,Z));
-        return 0.1*result;
-    }
-    double operator()(double R, double Z, double phi) const {
-        return this->operator()(R,Z);
-    }
-    double dR( double R, double Z)const
-    {
-        double psip = c_.psip(R,Z), psipR = c_.psipR(R,Z), theta_ = k_*theta(R,Z);
-        double result = (2.*psip*psipR - (psi0_+psi1_)*psipR)*cos(theta_) 
-            - (psip-psi0_)*(psip-psi1_)*sin(theta_)*k_*thetaR(R,Z);
-        return 0.1*result;
-    }
-    double dRR( double R, double Z)const
-    {
-        double psip = c_.psip(R,Z), psipR = c_.psipR(R,Z), theta_=k_*theta(R,Z), thetaR_=k_*thetaR(R,Z);
-        double psipRR = c_.psipRR(R,Z);
-        double result = (2.*(psipR*psipR + psip*psipRR) - (psi0_+psi1_)*psipRR)*cos(theta_)
-            - 2.*(2.*psip*psipR-(psi0_+psi1_)*psipR)*sin(theta_)*thetaR_
-            - (psip-psi0_)*(psip-psi1_)*(k_*thetaRR(R,Z)*sin(theta_)+cos(theta_)*thetaR_*thetaR_);
-        return 0.1*result;
-            
-    }
-    double dZ( double R, double Z)const
-    {
-        double psip = c_.psip(R,Z), psipZ = c_.psipZ(R,Z), theta_=k_*theta(R,Z);
-        double result = (2*psip*psipZ - (psi0_+psi1_)*psipZ)*cos(theta_) 
-            - (psip-psi0_)*(psip-psi1_)*sin(theta_)*k_*thetaZ(R,Z);
-        return 0.1*result;
-    }
-    double dZZ( double R, double Z)const
-    {
-        double psip = c_.psip(R,Z), psipZ = c_.psipZ(R,Z), theta_=k_*theta(R,Z), thetaZ_=k_*thetaZ(R,Z);
-        double psipZZ = c_.psipZZ(R,Z);
-        double result = (2.*(psipZ*psipZ + psip*psipZZ) - (psi0_+psi1_)*psipZZ)*cos(theta_)
-            - 2.*(2.*psip*psipZ-(psi0_+psi1_)*psipZ)*sin(theta_)*thetaZ_
-            - (psip-psi0_)*(psip-psi1_)*(k_*thetaZZ(R,Z)*sin(theta_) + cos(theta_)*thetaZ_*thetaZ_ );
-        return 0.1*result;
-    }
-    private:
-    double theta( double R, double Z) const {
-        double dR = R-R_0_;
-        if( Z >= 0)
-            return acos( dR/sqrt( dR*dR + Z*Z));
-        else
-            return 2.*M_PI-acos( dR/sqrt( dR*dR + Z*Z));
-    }
-    double thetaR( double R, double Z) const {
-        double dR = R-R_0_;
-        return -Z/(dR*dR+Z*Z);
-    }
-    double thetaZ( double R, double Z) const {
-        double dR = R-R_0_;
-        return dR/(dR*dR+Z*Z);
-    }
-    double thetaRR( double R, double Z) const {
-        double dR = R-R_0_;
-        return 2*Z*dR/(dR*dR+Z*Z)/(dR*dR+Z*Z);
-    }
-    double thetaZZ( double R, double Z) const { return -thetaRR(R,Z);}
-    double R_0_;
-    double psi0_, psi1_, k_;
-    const Collective& c_;
-};
-
-//takes the magnetic field as chi
-template<class Collective>
-struct EllipticDirPerM
-{
-    EllipticDirPerM( const Collective& c, double R0, double psi_0, double psi_1, double k): func_(c, R0, psi_0, psi_1, k), bmod_(c, R0), br_(c, R0), bz_(c, R0) {}
-    double operator()(double R, double Z, double phi) const {
-        return this->operator()(R,Z);}
-    double operator()(double R, double Z) const {
-        double bmod = bmod_(R,Z), br = br_(R,Z), bz = bz_(R,Z);
-        return -(br*func_.dR(R,Z) + bz*func_.dZ(R,Z) + bmod*(func_.dRR(R,Z) + func_.dZZ(R,Z) ));
-
-    }
-    private:
-    FuncDirPer<Collective> func_;
-    Bmodule<Collective> bmod_;
-    BR<Collective> br_;
-    BZ<Collective> bz_;
-};
-
-//Blob function
-template<class Collective>
-struct FuncDirNeu
-{
-    FuncDirNeu( const Collective& c, double psi_0, double psi_1, double R_blob, double Z_blob, double sigma_blob, double amp_blob):
-        psi0_(psi_0), psi1_(psi_1), 
-        cauchy_(R_blob, Z_blob, sigma_blob, sigma_blob, amp_blob){}
-
-    double operator()(double R, double Z, double phi) const {
-        return this->operator()(R,Z);}
-    double operator()(double R, double Z) const {
-        return cauchy_(R,Z);
-        //double psip = psip_(R,Z);
-        //return (psip-psi0_)*(psip-psi1_)+cauchy_(R,Z);
-        //return (psip-psi0_)*(psip-psi1_);
-    }
-    double dR( double R, double Z)const
-    {
-        return cauchy_.dx(R,Z);
-        //double psip = psip_(R,Z), psipR = psipR_(R,Z);
-        //return (2.*psip-psi0_-psi1_)*psipR + cauchy_.dx(R,Z);
-        //return (2.*psip-psi0_-psi1_)*psipR;
-    }
-    double dRR( double R, double Z)const
-    {
-        return cauchy_.dxx(R,Z);
-        //double psip = psip_(R,Z), psipR = psipR_(R,Z);
-        //double psipRR = psipRR_(R,Z);
-        //return (2.*(psipR*psipR + psip*psipRR) - (psi0_+psi1_)*psipRR)+cauchy_.dxx(R,Z);
-        //return (2.*(psipR*psipR + psip*psipRR) - (psi0_+psi1_)*psipRR);
-            
-    }
-    double dZ( double R, double Z)const
-    {
-        return cauchy_.dy(R,Z);
-        //double psip = psip_(R,Z), psipZ = psipZ_(R,Z);
-        //return (2*psip-psi0_-psi1_)*psipZ+cauchy_.dy(R,Z);
-        //return (2*psip-psi0_-psi1_)*psipZ;
-    }
-    double dZZ( double R, double Z)const
-    {
-        return cauchy_.dyy(R,Z);
-        //double psip = psip_(R,Z), psipZ = psipZ_(R,Z);
-        //double psipZZ = psipZZ_(R,Z);
-        //return (2.*(psipZ*psipZ + psip*psipZZ) - (psi0_+psi1_)*psipZZ)+cauchy_.dyy(R,Z);
-        //return (2.*(psipZ*psipZ + psip*psipZZ) - (psi0_+psi1_)*psipZZ);
-    }
-    private:
-    double psi0_, psi1_;
-    dg::Cauchy cauchy_;
-};
-
-
-//takes the magnetic field multiplied by (1+0.5sin(theta)) as chi
-template<class Collective>
-struct BmodTheta
-{
-    BmodTheta( const Collective& c, double R0): R_0_(R0), bmod_(c, R0){}
-    double operator()(double R,double Z, double phi) const{
-        return this->operator()(R,Z);}
-    double operator()(double R,double Z) const{
-        return bmod_(R,Z)*(1.+0.5*sin(theta(R,Z)));
-    }
-    private:
-    double theta( double R, double Z) const {
-        double dR = R-R_0_;
-        if( Z >= 0)
-            return acos( dR/sqrt( dR*dR + Z*Z));
-        else
-            return 2.*M_PI-acos( dR/sqrt( dR*dR + Z*Z));
-    }
-    double R_0_;
-    Bmodule<Collective> bmod_;
-
-};
-
-//take BmodTheta as chi
-template<class Collective>
-struct EllipticDirNeuM
-{
-    EllipticDirNeuM( const Collective& c, double R0, double psi_0, double psi_1, double R_blob, double Z_blob, double sigma_blob, double amp_blob): R_0_(R0), 
-    func_(c, psi_0, psi_1, R_blob, Z_blob, sigma_blob,amp_blob), bmod_(c, R0), br_(c, R0), bz_(c, R0) {}
-    double operator()(double R, double Z) const {
-        double bmod = bmod_(R,Z), br = br_(R,Z), bz = bz_(R,Z), theta_ = theta(R,Z);
-        double chi = bmod*(1.+0.5*sin(theta_));
-        double chiR = br*(1.+0.5*sin(theta_)) + bmod*0.5*cos(theta_)*thetaR(R,Z);
-        double chiZ = bz*(1.+0.5*sin(theta_)) + bmod*0.5*cos(theta_)*thetaZ(R,Z);
-        return -(chiR*func_.dR(R,Z) + chiZ*func_.dZ(R,Z) + chi*( func_.dRR(R,Z) + func_.dZZ(R,Z) ));
-
-    }
-    double operator()(double R, double Z, double phi) const {
-        return this->operator()(R,Z);
-    }
-    private:
-    double theta( double R, double Z) const {
-        double dR = R-R_0_;
-        if( Z >= 0)
-            return acos( dR/sqrt( dR*dR + Z*Z));
-        else
-            return 2.*M_PI-acos( dR/sqrt( dR*dR + Z*Z));
-    }
-    double thetaR( double R, double Z) const {
-        double dR = R-R_0_;
-        return -Z/(dR*dR+Z*Z);
-    }
-    double thetaZ( double R, double Z) const {
-        double dR = R-R_0_;
-        return dR/(dR*dR+Z*Z);
-    }
-    double R_0_;
-    FuncDirNeu<Collective> func_;
-    Bmodule<Collective> bmod_;
-    BR<Collective> br_;
-    BZ<Collective> bz_;
-};
-
-//the psi surfaces
-template<class Collective>
-struct FuncXDirNeu
-{
-    FuncXDirNeu( const Collective& c, double psi_0, double psi_1):
-        c_(c), psi0_(psi_0), psi1_(psi_1){}
-
-    double operator()(double R, double Z, double phi) const {
-        return this->operator()(R,Z);}
-    double operator()(double R, double Z) const {
-        double psip = c_.psip(R,Z);
-        return (psip-psi0_)*(psip-psi1_);
-    }
-    double dR( double R, double Z)const
-    {
-        double psip = c_.psip(R,Z), psipR = c_.psipR(R,Z);
-        return (2.*psip-psi0_-psi1_)*psipR;
-    }
-    double dRR( double R, double Z)const
-    {
-        double psip = c_.psip(R,Z), psipR = c_.psipR(R,Z);
-        double psipRR = c_.psipRR(R,Z);
-        return (2.*(psipR*psipR + psip*psipRR) - (psi0_+psi1_)*psipRR);
-            
-    }
-    double dZ( double R, double Z)const
-    {
-        double psip = c_.psip(R,Z), psipZ = c_.psipZ(R,Z);
-        return (2*psip-psi0_-psi1_)*psipZ;
-    }
-    double dZZ( double R, double Z)const
-    {
-        double psip = c_.psip(R,Z), psipZ = c_.psipZ(R,Z);
-        double psipZZ = c_.psipZZ(R,Z);
-        return (2.*(psipZ*psipZ + psip*psipZZ) - (psi0_+psi1_)*psipZZ);
-    }
-    private:
-    Collective c_;
-    double psi0_, psi1_;
-};
-
-//take Bmod as chi
-template<class Collective>
-struct EllipticXDirNeuM
-{
-    EllipticXDirNeuM( const Collective& c, double R0, double psi_0, double psi_1): R_0_(R0), 
-    func_(c, psi_0, psi_1), bmod_(c, R0), br_(c, R0), bz_(c, R0) {}
-    double operator()(double R, double Z) const {
-        double bmod = bmod_(R,Z), br = br_(R,Z), bz = bz_(R,Z);
-        double chi = 1e4+bmod; //bmod can be zero for a Taylor state(!)
-        //double chi = bmod; //bmod can be zero for a Taylor state(!)
-        double chiR = br;
-        double chiZ = bz;
-        return -(chiR*func_.dR(R,Z) + chiZ*func_.dZ(R,Z) + chi*( func_.dRR(R,Z) + func_.dZZ(R,Z) ));
-        //return -( func_.dRR(R,Z) + func_.dZZ(R,Z) );
-
-    }
-    double operator()(double R, double Z, double phi) const {
-        return this->operator()(R,Z);
-    }
-    private:
-    double R_0_;
-    FuncXDirNeu<Collective> func_;
-    Bmodule<Collective> bmod_;
-    BR<Collective> br_;
-    BZ<Collective> bz_;
-};
-
-//take Blob and chi=1
-template<class Collective>
-struct EllipticBlobDirNeuM
-{
-    EllipticBlobDirNeuM( const Collective& c, double psi_0, double psi_1, double R_blob, double Z_blob, double sigma_blob, double amp_blob): 
-    func_(c, psi_0, psi_1, R_blob, Z_blob, sigma_blob, amp_blob){}
-    double operator()(double R, double Z) const {
-        return -( func_.dRR(R,Z) + func_.dZZ(R,Z) );
-    }
-    double operator()(double R, double Z, double phi) const {
-        return this->operator()(R,Z);
-    }
-    private:
-    double R_0_;
-    FuncDirNeu<Collective> func_;
-};
-
-template<class Collective>
-struct EllipticDirSimpleM
-{
-    EllipticDirSimpleM( const Collective& c, double psi_0, double psi_1, double R_blob, double Z_blob, double sigma_blob, double amp_blob): func_(c, psi_0, psi_1, R_blob, Z_blob, sigma_blob, amp_blob) {}
-    double operator()(double R, double Z, double phi) const {
-        return -(( 1./R*func_.dR(R,Z) + func_.dRR(R,Z) + func_.dZZ(R,Z) ));
-
-    }
-    private:
-    FuncDirNeu<Collective> func_;
-};
-
-///@} 
-} //namespace solovev
